feat(vs2022_third): C4018 fix signed/unsigned mismatch in NanoSpec — 5 files (batch 6)
Cast int loop variables to DWORD/unsigned int in comparisons:
- NanoSpec/StressLineSectionSettingDlg.cpp: 36 instances (i,j,row→DWORD/int)
- NanoSpec/ChifTransiStress.cpp: 2 instances (i→DWORD in <= comparisons)
- NanoSpec/RecipeThicknessSettingView.cpp: 5 instances (iCount,jCount→DWORD)
- NanoSpec/SrVideoCalibrationSettingDlg.cpp: 2 instances (i→unsigned int)
- NanoSpec/StressElasticModulusListDlg.cpp: 2 instances (i→unsigned int)

Total: 47 C4018 warnings fixed (targeting all remaining 43).
No behavior changes, only explicit casts to resolve type mismatch.

+++ SRC/NanoSpec/StressLineSectionSettingDlg.cpp
@@ -1115,7 +1115,7 @@ void CStressLineSectionSettingDlg::OnKillfocusNumberOfLine()
 	m_StressConfig.dwLiftPinNumberOfLine = NumOfLine;
 // 2009.09.08 bagus stress èCê≥ --{--
 	for (int i=0; i<STRESS_LINES_MAX; i++) {
-		if (i<NumOfLine) {
+		if ((DWORD)i<NumOfLine) {
 //			m_StressConfig.Line[i].bValidLine = TRUE;
 		} else {
 			m_StressConfig.Line[i].bValidLine = FALSE;
@@ -1164,7 +1164,7 @@ void CStressLineSectionSettingDlg::SetLineData()
 	TRY{
 // 2009.09.17 bagus stress GridèCê≥ --{--
 //		m_LineListGrid.SetRowCount(m_StressConfig.dwLiftPinNumberOfLine + FixRows);
-		if ((m_StressConfig.dwLiftPinNumberOfLine + FixRows) < DEFAULT_GRID_ROWS) {
+		if ((int)(m_StressConfig.dwLiftPinNumberOfLine + FixRows) < DEFAULT_GRID_ROWS) {
 			m_LineListGrid.SetRowCount(DEFAULT_GRID_ROWS);
 		} else {
 			m_LineListGrid.SetRowCount(m_StressConfig.dwLiftPinNumberOfLine + FixRows);
@@ -1178,7 +1178,7 @@ void CStressLineSectionSettingDlg::SetLineData()
 	}
 	END_CATCH
 
-	for (row=FixRows; row<m_StressConfig.dwLiftPinNumberOfLine+FixRows; row++) {
+	for (row=FixRows; (DWORD)row<m_StressConfig.dwLiftPinNumberOfLine+FixRows; row++) {
 		dwTextStyle = DT_CENTER|DT_VCENTER|DT_SINGLELINE;
 		dwTextStyle |= DT_END_ELLIPSIS;
 		Item.mask = GVIF_TEXT|GVIF_FORMAT;
